requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-apache-spark==4.9.1
pyspark==3.5.0
pandas==2.1.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
